Download packages from GitLab as tarballs
When a URL is recognized to be GitLab, we can use their API to download
a tarball like we do with GitHub. This avoids having to download entire
Git repos for GitLab-hosted packages.

Note however that self-hosted GitLab instances don't necessarily have to
have "gitlab" in their URLs, but those that do typically start with
"gitlab," so it provides a reasonable heuristic.