Use `create_artifact()` inside of `download_artifact()`
This one was subtle: if a user listed the wrong git-tree-sha1 within their
`Artifacts.toml`, if it was not already installed, we would fill that
directory up with the wrong contents, causing all sorts of confusion
later on.  This better verifies that the directories on-disk have
exactly the contents that they advertise, and helps users to realize
when they have put the wrong hash in their `Artifacts.toml` files.