Don't verify tree hash correctness on Windows, because we're permissions-blind
We dodge the inability for Julia to use `chmod()` and `stat()` to
set/get executable permissions on Windows, solving our git tree hashing
issues temporarily.  Eventually we will want to augment `stat()` and
`chmod()` to actually probe the relevant ACLs on Windows to get those
values (potentially gated behind a kwarg flag, as I hear this can be
expensive).