Fix `compress_versions()` confusion with build number VersionNumber's
This is not a permanent fix, it's to paper over the fact that we have
two, slightly incompatible concepts of a `VersionNumber` within `Pkg`;
the `VersionNumber` and the `VersionBound`.  The latter does not support
`build` or `prerelease` specifiers, which causes much confusion within
`compress_versions()` when it attempts to compare equality of
`VersionNumber` objects with their supposedly-equivalent `VersionBound`
objects.  To fix this, we explcitly strip those pieces of information
before doing any comparisons.