Improve `@artifact_str` relocatability
By moving some work to compile-time, we are able to avoid usage of
`@__DIR__` and even parsing of `Artifacts.toml` files at all during
module `__init__()`; this means that relocated `.ji` files should
maintain all knowledge of artifacts without needing to hit baked
filesystem paths, which is a definite plus.  Error messages will
continue to point to the wrong location, but that is forgivable.

This may speed up JLL package load time by a small amount, however we
still need to hit the disk for things like `Overrides.toml`, and we
still need to `dlopen()` things, etc... so I do not expect this to have
a large impact.