Add precompile section for `@artifact_str`
This adds a precompile section so that we can avoid precompiling `@artifact_str` so much.
We generate a simple artifact, bind it out to an `Artifacts.toml` file, then read it in
again with an `include()`.

This also adds a rudimentary `precompile_script` test, but it doesn't
actually fail if the script has an error in it, so it definitely needs
improvement.