Ignore symlinks when fixing permissions during artifacts tests (#1721)
When doing Artifact tests, we need to be resilient to strange umasks
that we have inherited from our environment, screwing up git tree hash
calculations.  To do this, we have a `walkdir()` -> `chmod()` wrapper
function that started setting permissions on what it thought were files
(due to the fixed `follow_symlink == false` behavior in
https://github.com/JuliaLang/julia/pull/35006) but were actually
directories.  This caused `chmod()` to apply `0o664` permissions to
directories, which were then inaccessible to `rm()` as it tried to
delete them.

This fixes https://github.com/JuliaLang/Pkg.jl/issues/1716