Reduce footprint for invalidation
This is a companion PR to https://github.com/JuliaLang/julia/pull/35714
and accounts for a subset of the gains reported there.
The more specific implementations of `==` and `hash` avoid a `collect`;
that branch prevents many `collect`-induced invalidations via the
`iteratorsize` trick, but this change is probably a good idea anyway.

The more important and subtle one is the one in TOML._print.
This is used by a large number of methods in Pkg, so it's important
to prevent it from being invalidated.
Do to nospecialize markers, this method gets inferred for partially-
specified types, and the `a` is typically just `AbstractDict`.
However, `Iterators.Pairs` is also an AbstractDict, and the return
type of `keys` cannot be inferred from the parameters of
`AbstractDict{K,V}` because the `itr` field of a pair uses a later
parameter.

This short-circuits the whole problem by "hand-inlining" the result
of `invoke(keys, Tuple{AbstractDict})` at the call site, i.e.,
just creating a `KeySet` directly.