Clean up `jl_XXX-download.gz.sha256` files
Because PlatformEngines assumes you may want to verify a tarball later, it spits out `.sha256` files to optimize for hashing in the future.  This is fine when you're downloading to a permanent location or a directory that was created by `mktempdir()`, but if your download location is just a file sitting in `/tmp` which you clean up manually afterward, you need to clean up that `.sha256` file as well.