telemetry: ~/.julia/servers/telemetry.toml for defaults
This allows, for example, a user to do

    echo 'telemetry = false' >> ~/.julia/servers/telemetry.toml

and opt out of all telemetry for all servers they might connect to.
If they only want to opt out of sending any telemetry that is based
on a client UUID, then they can do

	echo 'client_uuid = false' >> ~/.julia/servers/telemetry.toml

This can be done anywhere in the depot path, so if a sysadmin wants
to set up a shared system so that all users default to sending no
telemetry, they can do that. A user can opt back into telemetry by
Putting an empty default `telemetry.toml` file in their own depot.