no-arg activate: if already deactivated, activate current project
Previously there was no convenient way to activate the current
project from the Pkg REPL or API. This changes `Pkg.activate()` so
that if there is no active project to deactivate, it instead
activates `Base.current_project()`. This means that we can now say
that `julia --project` is exactly equivlent to `julia` followed by
`pkg> activate`.