correct documentation of default value for JULIA_PKG_DEVDIR
The default for `JULIA_PKG_DEVDIR` is *not* `.julia/dev` but `joinpath(DEPOT_PATH[1],"dev")`. This is important: users are currently not warned in [the documentation of `JULIA_DEPOT_PATH`](https://docs.julialang.org/en/v1/manual/environment-variables/#JULIA_DEPOT_PATH) that changing the first element of that environment variable will change the default of `JULIA_PKG_DEVDIR` as well.