Clarify activate command signatures
The signatures in the help for the activate command were confusing. It seemed to imply that the `activate --shared` and `activate --temp` each took an optional path, when in fact `activate --shared` requires a path and `activate --temp` does not accept a path. This clarifies the command signatures for these two cases.