[BinaryPlatforms]: fix triplet breakage
We momentarily broke the MacOS (and FreeBSD, although I'm not sure that
was actually noticed ;P) ecosystem by changing slightly the rules around
`triplet(::Platform)`.  `Base.BinaryPlatforms` keeps track of the
`os_version` and appends it to the end of the triplet for `MacOS` and
`FreeBSD`, but on older versions of julia they are always expected to be
`14` and `11.1`, respectively.  This PR maintains that willful ignorance
when an old package calls `Pkg.BinaryPlatforms.triplet()`, but for new
packages that are ready to wake up and stop living in a simulation, they
can use `Base.BinaryPlatforms.triplet()` to get the full, glorious,
properly-versioned triplet string.