send Pkg server value itself as a header to the Pkg server (#2134)
This is useful because there may be redirection going on and it may
not be clear how someone got to a particular pkg server. Setting
this header is not a risk even though it captures the content of
the `JULIA_PKG_SERVER` environment variable, because if the content
of the variable happens to be something sensitive instead of the
address of a package server, then the request will never reach that
package server in the first place, let alone the header.