Windows: prefer git cloning registries (close #2014)
Since untarring a tarball like the registry is so slow on Windows,
we prefer to git clone registries instead. Git is slow on Windows
as well, but git updates are faster since git only updates the
files that need to be modified, which is much faster than unpacking
a whole new registry.