Do not set links as read-only with `set_readonly`
`set_readonly` is supposed to set only files as read-only, but in practice if
there is a link to a directory it'll set the target directory as read-only,
causing the files inside it to be unremovable without first fixing permissions
of the directory.

This prevents `set_readonly` from changing the mode of any link.