Do not set links as read-only with `set_readonly` (#2286)
`set_readonly` is supposed to set only files as read-only, but in practice if
there is a link to a directory it'll set the target directory as read-only,
causing the files inside it to be unremovable without first fixing permissions
of the directory.

This prevents `set_readonly` from changing the mode of any link.

(cherry picked from commit 731633dc0ca3db200a9e23d0cdc5dfa40e56176d)