[ext/HSG]: Store next release _and_ latest nightly (#2418)
In the event that we are in an RC phase where we have three "latest"
branch tips (e.g. latest stable: 1.5.3, latest RC: 1.6.0-rc1 and latest
`master`: 1.7.0-DEV) each with their own set of standard libraries, we
must store this information in Pkg so that we pass tests when running on
latest `master`. It is harmless for this information to be baked into an
older Julia, as we make no guarantees that Julia 1.6.0 final should have
perfect information about what standard libraries 1.7.0 will have.

(cherry picked from commit 7a9d96545a9c30009b6653374a0ee960383441b6)