Cache registry data for PkgServer registries over Pkg invocations
There was already a cache in place that cached parsing of TOML files.
However, many other parts of the regsitry handling (like uncompressing the version ranges) was not.
In the future we want to move to keeping the registry in its tarred form which means that re-reading
it will be expensive, so let's cache it instead.

This takes for example the time to add Plots over and over from 234ms to 150ms.