Automatically call `Pkg.gc()` after `up`, `pin`, `free` and `rm`
We want to call `Pkg.gc()` every now and then, so let's automatically
call it after operations that can themselves result in orphaned package
versions (e.g. `rm` or `up`).  We wait the default GC collect interval,
using `mtime(orphanage_path)` as our marker of the last time we have
collected.  After each of the above operations, we query the current
time as compared to the last known GC time, hitting the filesystem only
when there is a possibility that we might need to GC.

Also, we now always write out an orphanage file (even if it's empty) to
ensure that the timestamp is always available.