Improve the local optimizer at the end of resolve
For each package, instead of trying to bump just one package,
try to propagate the effect of the bump outward through the graph.

Simple example: if package Av1 depends on Bv1 and Av2 depends on Bv2,
and we are considering bumping Av1 to Av2, with the previous code we
would just give up. With this change we try to bump Bv1 to Bv2 and see
if that works.

Still todo: polishing, tests, improve logging (?)