Fix a bug, further improve the optimizer, add test
Bugfix: in propagate_constraints, actually propagate the constraints
throughout the reachable graph. Otherwise, inconsistent situations might
arise (under peculiar circumstances).

Optimizer improvements: allow the local optimizer to also install new
packages, if this allows other packages to be bumped.

The test is a reduced form that fails with the prvious optimizer (even
with the bug fix) and succeeds with the new one.