Fix REPL tests lacking `stdin` on `Distributed` workers
When running the `Pkg` tests as part of the Julia test suite (e.g. via
`Base.runtests(["triplequote", "Pkg"])`, note that you need multiple
tests in order to trigger the `Distributed` behavior) we get run inside
of a `Distributed` worker which does not have access to `stdin`.
Attempting to run `redirect_stdin()` then fails with an error from
`dup()` such as the following:

```
SystemError: dup: Bad file descriptor
Stacktrace:
  [1] systemerror(p::String, errno::Int32; extrainfo::Nothing)
    @ Base ./error.jl:176
  [2] #systemerror#80
    @ ./error.jl:175 [inlined]
  [3] systemerror
    @ ./error.jl:175 [inlined]
  [4] dup
    @ ./libc.jl:37 [inlined]
  [5] _redirect_io_libc
    @ ./stream.jl:1213 [inlined]
  [6] (::Base.RedirectStdStream)(handle::Base.PipeEndpoint)
    @ Base ./stream.jl:1223
```

To fix this we follow the same convention as used in the base tests [0,1]
That is, we schedule work items on worker 1.  The main difference
between this and the base tests is that we must import `Pkg` and its
test utils on worker 1.

[0] https://github.com/JuliaLang/julia/blob/ef835ac1337ccc1463934af092ab37d9903eb6e3/test/spawn.jl#L336-L341
[1] https://github.com/JuliaLang/julia/blob/ef835ac1337ccc1463934af092ab37d9903eb6e3/test/show.jl#L723-L731