Update environments.md
I think the docs are not completely clear on what happens if one just `activate`s an environment but does not instantiate it. 
In this case Julia will throw an error if one attempts to `import` a package that is not already installed matching the precise version listed in the manifest file, if present.
This should be stated explicitly.