Clarify what happens if one activates a project without instantiating it (#3100)
* Update environments.md

I think the docs are not completely clear on what happens if one just `activate`s an environment but does not instantiate it.
In this case Julia will throw an error if one attempts to `import` a package that is not already installed matching the precise version listed in the manifest file, if present.
This should be stated explicitly.

* Update docs/src/environments.md

Co-authored-by: Ian Butterworth <i.r.butterworth@gmail.com>

* Update docs/src/environments.md

Co-authored-by: Ian Butterworth <i.r.butterworth@gmail.com>

* Update docs/src/environments.md

* Update docs/src/environments.md

Co-authored-by: Ian Butterworth <i.r.butterworth@gmail.com>
(cherry picked from commit c82cabde9c02c3ab423c40d5a38d74453876aa8e)