Respect manifest if provided in Project.toml
I recently discovered that loading in Base supports specifying
an alternative `Manifest.toml` in a `Project.toml`. While playing
around with this though, I discovered that Pkg.jl doesn't play as
nice with this. It supports the parsing/storing of it, but gets
confused often because so many path calculations are made relative
to the _project file_ instead of the now-different manifest file.

This is relevant when there are devv-ed packages in the specified
Manifest and the relative path calculations should be made wrt to
the _manifest_ not the _project_.

The changes here, hence, are mainly about auditing uses of
`env.project_file` and changing them to `env.manifest_file` if
necessary. The cases where `manifest_file` should be used, as far
as I understand, are where the `PackageSpec` we're computing a
path for came from a manifest entry instead of a project dep or elsewhere.