Fix a minor manifest file detection issue (#3356)
This is due to an oversight on my part in https://github.com/JuliaLang/Pkg.jl/pull/3303
where I didn't realize you could be in a different directory than the currently activated
environment. In such a setup, and if the project manually specified a manifest to use,
we weren't getting the manifest file path _relative_ to the project directory if the
provided path was indeed relative. The fix here is to first check if the manually provided
manifest path is already abspath, and if not, ensure we get the abspath _relative_ to the
the project directory.

(cherry picked from commit 3f742c38c694b9049198283cfc234ff9ee772fed)