Ignore lack of project.toml for removed stdlibs when `julia_version != VERSION`
Because DelimitedFiles was removed, the assumption in this conditional
(that if something is a stdlib, it exists locally) has been broken.
Because I don't think it's reasonable to go out and clone stdlibs at
this point, I'm just skipping the dependency (And any of its
dependencies) because this codepath is primarily used by BinaryBuilder,
where we really only care about JLLs anyway, which are never
unregistered.