Verify git-tree-sha1 of registry downloads
This patch adds verification of registries downloaded from Pkg server
when using the default method of keeping registries as compressed
tarballs. This verification is done before the registry is moved in
place to ensure that a Pkg server that serves bad content does not also
poison the users local machine.