do not silently accept broken `julia` compat
Pkg now raises an error if the julia compat is not met for a package. `ENV["JULIA_PKG_CHECK_JULIA_COMPAT"]` can be used to override this check.

This change was made due to new developers getting confused why their julia 1.8 does not work with a dev-ed package that has a julia compat set to 1.9.