Ensure that `string(::VersionSpec)` is compatible with `semver_spec()`
It is an unfortunate coincidence that if you try to print out a
`VersionSpec` to a project's `compat` field, it can be in a format that
`semver_spec()` rejects due to the ambiguity of `1.2.0-1.4.0` being
either a prerelease with a very strange prerelease tag, or a range of
versions.  We reduce the ambiguity by changing `VersionSpec` to always
print its bounds with a separating space.