Fix lacking `subdir` information when altering packages (#3597)
If a package we're altering is already in the manifest, we try to load
some information from the manifest, but we forgot to load `subdir` in
these two cases.  Add tests to ensure that `add`ing or `dev`ing the same
package twice in the same environment does not error.

This fixes #3391.