Make Pkg more resilient to reading older manifests where ex-stdlibs are listed (#3634)
* Make Pkg more resilient to reading older manifests where ex-stdlibs are listed

This should prevent errors like:

```
ERROR: Could not locate the source code for the DelimitedFiles package. Are you trying to use a manifest generated by a different version of Julia?
```

When using older manifests that list DelimitedFiles in the stdlib format even though DelimitedFiles now is upgradable and has e.g. a `git-tree-sha1` entry when generated with current Pkg