RFC: stop testing non-ARM on mac on CI (#3794)
* RFC: stop testing non-ARM on mac on CI

It feels a little bit excessive to me to test both arm and intel on mac since there isn't really many code paths in Julia (and especially Pkg) that are architecture dependent like this

It also feels excessive to test all builds with pkg.julialang.org and not. Should be enough with just one linux that tests both.