Add workspace feature (#3841)
This feature allows one to define a "workspace" in a project file by adding

```toml
[workspace]
projects = ["packageA", "docs", "test"]
```

to it. All projects in a workspace are resolved together into a single manifest file in the "root" project (the project that is not included by any other project in a workspace). The compat of all dependencies are intersected among the projects.

Currently, every project in a workspace must be in a folder at the same level as the project that includes it into the workspace.

If the test folder is included in the workspace, running `Pkg.test` is similar to just activating the test project and including the `runtest.jl` file.

Many functions now take the `workspace` argument to decide if the function should operate on the active project or on the full workspace:
- `Pkg.precompile`
- `Pkg.status`
- `Pkg.why`
- `Pkg.instantiate`