use the correct TOML.Parser coupled to Base.TOMLCache (#3921)
This assumption will be broken by #54739 and not recoverable, so we need
to first fix this Pkg bug. Since print is only in TOML (the stdlib),
this still needs the TOML package as well. Since some of the helpers are
only in TOML, this only changes the places that use Base.TOMLCache
features directly (not upgradable) with added Dates support, but uses
TOML for everything else.

Fixes JuliaLang/julia#53196