Make Pkg more robust to poor inference
A few things in Pkg rely on inference to figure out (empty)
collection types and error if inference can't figure it because
the subsequent method signatures are narrow.
I wouldn't really care for any other package, but when working
on inference it's not uncommon for it to temporarily be somewhat
broken. It's useful to still have Pkg available so that the build
finishes and that it is possible to install extra packages for
debugging.