Fix artifact directories not having traversal permissions
It turns out that Windows requires the executable bit set if the
`BYPASS_TRAVERSE_CHECKING` privilege is not attached to a user's
account.  It's simply more correct to ensure that our directories have
this bit set, and unfortunately our `filemode()` function call is
not complete on Windows and does not include this bit, so we manually
add it in on Windows.