Doc section for Pkg and Storage Protocols
This adds a section in the docs with the Package and Storage protocols. A collation of these:

1. Pkg & Storage Protocols [https://github.com/JuliaLang/Pkg.jl/issues/1377](https://github.com/JuliaLang/Pkg.jl/issues/1377)
2. Authenticated Pkg Client Support: [https://github.com/JuliaLang/Pkg.jl/pull/1538](https://github.com/JuliaLang/Pkg.jl/pull/1538)
3. Authentication Hooks: [https://github.com/JuliaLang/Pkg.jl/pull/1630](https://github.com/JuliaLang/Pkg.jl/pull/1630)

Mostly retained the original text, but updated with changes where applicable. Hoping that this will serve as a useful single point of reference. It can sometimes get confusing while navigating the github PRs and comments to figure out the final state. This can be updated if and when things change in future.