Fix some tests when run in base
Part of https://github.com/JuliaLang/julia/issues/60058. When we don't
have a log file setup, the default IO is set to devnull. However, this
causes `Pkg.status` to early out and skip some validity checks that
we are testing for. Fix this by passing an appropriate io directly
to the API. For the REPL test, which don't have an IO argument,
override the global default IO, but to make that more robust also
switch it to a ScopedValue (which Pkg is already using elsewhere).