Isolate more tests
As seen in https://github.com/JuliaLang/julia/pull/60197, the test I
added to make sure that `make test-Pkg` doesn't corrupt the builtin
depot is firing properly - we had several more tests that were
improperly isolated. A related issue is that if `Preferences` is
not in the depot that starts the test we were trying to Pkg.add
it, causing the same problem. Fix that by treating it like the
other external packages (Aqua, HistoricalStdlibVersions) and
making the corresponding tests optional.